feat: update getUserDetails method to use int for user ID and return user details with associated orders

+++ backend/bikePartsHub/src/main/java/com/bphTeam/bikePartsHub/service/UserService.java
@@ -4,5 +4,5 @@ import com.bphTeam.bikePartsHub.dto.response.UserResponseDto;
 
 public interface UserService {
 
-    UserResponseDto getUserDetails(Long userId);
+    UserResponseDto getUserDetails(int userId);
 }
